Pair coding session outcomes, moved things about a bit, added some classes from useTab.
