--- src/components/Tab/tab.st.css
@@ -1,5 +1,0 @@
-/* tab.st.css CSS API */
-.root {
-  margin-right: 10px;
-  padding: 10px;
-}
