--- src/components/TabPanel/tabPanel.st.css
@@ -1,4 +1,0 @@
-/* tabPanel.st.css CSS API */
-.root {
-  padding: 10px;
-}
