+++ src/components/Tabs/Tab.tsx
@@ -1,10 +1,9 @@
 import { Ref, useRef, ReactElement, forwardRef } from "react";
-import { useTab } from "react-aria";
+import { useTab, useFocusRing } from "react-aria";
 import { Node } from "@react-types/shared";
 import { TabListState } from "react-stately";
-
-import { st, classes } from "./tab.st.css";
-
+import { mergeProps } from "@react-aria/utils";
+import { st, classes } from "./tabs.st.css";
 export interface TabProps<T> {
   /** Add a class to the content div. */
   className?: string;
@@ -24,12 +23,27 @@ function Tab<T extends object>({
 }: TabProps<T>) {
   const { key, rendered } = item;
   const ref = useRef<HTMLDivElement>(null);
-  const { tabProps } = useTab({ key }, state, ref);
+  const { tabProps, isSelected, isDisabled, isPressed } = useTab(
+    { key },
+    state,
+    ref
+  );
+
+  const { isFocusVisible, focusProps } = useFocusRing();
 
   return (
     <div
-      className={st(classes.root, className)}
-      {...tabProps}
+      className={st(
+        classes.tab,
+        {
+          isSelected,
+          isDisabled,
+          isPressed,
+          isFocusVisible,
+        },
+        className
+      )}
+      {...mergeProps(tabProps, focusProps)}
       ref={ref}
       data-id={dataId}
     >
@@ -37,8 +51,6 @@ function Tab<T extends object>({
     </div>
   );
 }
-
-export default Tab;
 
 // forwardRef doesn't support generic parameters -> cast to the correct type.
 // https://stackoverflow.com/questions/58469229/react-with-typescript-generics-while-using-react-forwardref
