+++ src/indexLib.ts
@@ -70,7 +70,7 @@ export * from "./components/TableRow/TableRow";
 export { default as TablePagination } from "./components/TablePagination/TablePagination";
 export * from "./components/TablePagination/TablePagination";
 export { default as Tabs } from "./components/Tabs/Tabs";
-export * from "./components/Tabs/Tabs";
+// export * from "./components/Tabs";
 export { default as Toolbar } from "./components/Toolbar/Toolbar";
 export * from "./components/Toolbar/Toolbar";
 export { default as VisuallyHidden } from "./components/VisuallyHidden/VisuallyHidden";
